engine: move the portforwardcontroller into its own package
